Updated react scripts, firebase and added attendanceCode
i'm using node.js v23.7.0, which uses a newer version of OpenSSL that's not compatible with the older version of react-scripts (v3.0.1) in the project, so I updated the react-scripts.

i also added a randomized 6-letter attendanceCode to event creation, which showed up in firebase when i tested it. it should only show up for admin users but since i'm waiting for confirmed admin access, i can't tell if that part works