can add custom attendance code, event questions, redirect to home
event creator can add custom attendance codes (or have a randomly generated one) AND can add questions (required or optional) for attendees to answer when signing in (in addition to attendance code). This is updated in the firebase db. User is now redirected to home page after signing into event